feat(examples): make helloworldserver zero-allocation
Per review on #2362, fold the zeroallocserver example into
helloworldserver instead of adding it alongside.

The handler now uses Response.AppendBody/AppendBodyString with the
Append* helpers and a pooled scratch buffer, plus AcquireCookie in place
of a zero-value Cookie, so it runs at 0 allocs/op. TestZeroAllocation
checks this, mirroring TestAllocationServeConn.

The -compress flag is kept; compression itself allocates, so the claim
covers the uncompressed path. The raw request dump is dropped.

Refs #1764
