+++ examples/README.md
@@ -1,5 +1,4 @@
 # Code examples
 
 * [HelloWorld server](helloworldserver)
-* [Zero-allocation server](zeroallocserver)
 * [Static file server](fileserver)
