+++ examples/helloworldserver/helloworldserver.go
@@ -1,9 +1,12 @@
+// Command helloworldserver displays various request info with 0 allocs/op.
 package main
 
 import (
 	"flag"
-	"fmt"
 	"log"
+	"net"
+	"strconv"
+	"sync"
 
 	"github.com/valyala/fasthttp"
 )
@@ -18,6 +21,8 @@ func main() {
 
 	h := requestHandler
 	if *compress {
+		// Note: compression itself allocates. The 0 allocs/op claim covers
+		// requestHandler, i.e. the uncompressed path.
 		h = fasthttp.CompressHandler(h)
 	}
 
@@ -26,30 +31,66 @@ func main() {
 	}
 }
 
+// scratchPool holds reusable buffers for formatting non-[]byte values.
+var scratchPool = sync.Pool{
+	New: func() any {
+		b := make([]byte, 0, 64)
+		return &b
+	},
+}
+
 func requestHandler(ctx *fasthttp.RequestCtx) {
-	fmt.Fprintf(ctx, "Hello, world!\n\n")
+	buf := scratchPool.Get().(*[]byte) //nolint:forcetypeassert
+	defer scratchPool.Put(buf)
 
-	fmt.Fprintf(ctx, "Request method is %q\n", ctx.Method())
-	fmt.Fprintf(ctx, "RequestURI is %q\n", ctx.RequestURI())
-	fmt.Fprintf(ctx, "Requested path is %q\n", ctx.Path())
-	fmt.Fprintf(ctx, "Host is %q\n", ctx.Host())
-	fmt.Fprintf(ctx, "Query string is %q\n", ctx.QueryArgs())
-	fmt.Fprintf(ctx, "User-Agent is %q\n", ctx.UserAgent())
-	fmt.Fprintf(ctx, "Connection has been established at %s\n", ctx.ConnTime())
-	fmt.Fprintf(ctx, "Request has been started at %s\n", ctx.Time())
-	fmt.Fprintf(ctx, "Serial request number for the current connection is %d\n", ctx.ConnRequestNum())
-	fmt.Fprintf(ctx, "Your ip is %q\n\n", ctx.RemoteIP())
+	resp := &ctx.Response
+	resp.AppendBodyString("Hello, world!\n\n")
 
-	fmt.Fprintf(ctx, "Raw request is:\n---CUT---\n%s\n---CUT---", &ctx.Request)
+	resp.AppendBodyString("Request method is \"")
+	resp.AppendBody(ctx.Method())
+	resp.AppendBodyString("\"\nRequestURI is \"")
+	resp.AppendBody(ctx.RequestURI())
+	resp.AppendBodyString("\"\nRequested path is \"")
+	resp.AppendBody(ctx.Path())
+	resp.AppendBodyString("\"\nHost is \"")
+	resp.AppendBody(ctx.Host())
+	resp.AppendBodyString("\"\nQuery string is \"")
+	resp.AppendBody(ctx.QueryArgs().AppendBytes((*buf)[:0]))
+	resp.AppendBodyString("\"\nUser-Agent is \"")
+	resp.AppendBody(ctx.UserAgent())
+
+	resp.AppendBodyString("\"\nConnection has been established at ")
+	resp.AppendBody(fasthttp.AppendHTTPDate((*buf)[:0], ctx.ConnTime()))
+	resp.AppendBodyString("\nRequest has been started at ")
+	resp.AppendBody(fasthttp.AppendHTTPDate((*buf)[:0], ctx.Time()))
+
+	resp.AppendBodyString("\nSerial request number for the current connection is ")
+	resp.AppendBody(strconv.AppendUint((*buf)[:0], ctx.ConnRequestNum(), 10))
+
+	resp.AppendBodyString("\nYour ip is \"")
+	resp.AppendBody(appendIP((*buf)[:0], ctx.RemoteIP()))
+	resp.AppendBodyString("\"\n\n")
 
 	ctx.SetContentType("text/plain; charset=utf8")
 
 	// Set arbitrary headers
-	ctx.Response.Header.Set("X-My-Header", "my-header-value")
+	resp.Header.Set("X-My-Header", "my-header-value")
 
-	// Set cookies
-	var c fasthttp.Cookie
+	// Set cookies. AcquireCookie avoids the allocation a zero-value Cookie
+	// would cause.
+	c := fasthttp.AcquireCookie()
+	defer fasthttp.ReleaseCookie(c)
 	c.SetKey("cookie-name")
 	c.SetValue("cookie-value")
-	ctx.Response.Header.SetCookie(&c)
+	resp.Header.SetCookie(c)
+}
+
+// appendIP appends ip to dst without allocating. AppendIPv4 only handles
+// IPv4, so IPv6 falls back to net.IP.AppendText.
+func appendIP(dst []byte, ip net.IP) []byte {
+	if ip.To4() != nil {
+		return fasthttp.AppendIPv4(dst, ip)
+	}
+	dst, _ = ip.AppendText(dst)
+	return dst
 }
