--- examples/zeroallocserver/Makefile
@@ -1,6 +1,0 @@
-zeroallocserver: clean
-	go get -u github.com/valyala/fasthttp
-	go build
-
-clean:
-	rm -f zeroallocserver
